feat(游戏中心): 添加游戏帮助功能及俄罗斯转盘超时机制
1. 新增游戏帮助功能，包括HTML模板和渲染逻辑
2. 为俄罗斯转盘游戏添加60秒超时自动中弹机制
3. 增加强制结束游戏命令
4. 优化游戏状态提示信息

+++ src/plugins/handler.py
@@ -1,7 +1,7 @@
 '''
 Date: 2025-02-18 13:34:16
 LastEditors: yhl yuhailong@thalys-tech.onaliyun.com
-LastEditTime: 2025-06-02 22:53:51
+LastEditTime: 2025-06-02 23:24:08
 FilePath: /team-bot/jx3-team-bot/src/plugins/handler.py
 '''
 # src/plugins/chat_plugin/handler.py
@@ -10,7 +10,7 @@ from nonebot.rule import to_me
 from nonebot.typing import T_State
 from nonebot.adapters.onebot.utils import highlight_rich_message
 from nonebot.adapters.onebot.v11 import MessageEvent, MessageSegment, GroupMessageEvent, Bot, Message,GroupMessageEvent
-from .html_generator import render_html,render_help
+from .html_generator import render_html,render_help,render_game_help
 from .render_image import generate_html_screenshot, html_to_image
 from .api import check_default_team_exists, check_enroll, check_member, clear_teams, close_team, del_member, enroll_member, team_info, team_list, create_team, update_team_default, update_team_name,move_member,team_info_by_id,del_member_by_name
 from ..utils.index import find_default_team, find_earliest_team, find_id_by_team_name, format_teams, get_code_by_name, get_info_by_id, path_to_base64, upload_image,render_team_template,generate_team_stats
@@ -652,11 +652,11 @@ async def handle_random_black(bot: Bot, event: GroupMessageEvent, state: T_State
     msg = f"恭喜【{lucky_member.get('role_name')}】成为本次黑本幸运儿！"
     await RandomBlack.finish(message=Message(msg))
 
-# # 帮助
-Help = on_regex(pattern=r'^帮助$',priority=1)
+# # 开团帮助
+Help = on_regex(pattern=r'^开团帮助$',priority=1)
 @Help.handle()
 async def handle_help(bot: Bot, event: GroupMessageEvent, state: T_State):
-    if not await check_command_enabled(bot, event, "帮助"):
+    if not await check_command_enabled(bot, event, "开团帮助"):
         return
     
     # 发送处理提示
@@ -671,6 +671,29 @@ async def handle_help(bot: Bot, event: GroupMessageEvent, state: T_State):
     
     # 发送图片
     await Help.finish(MessageSegment.image(path_to_base64(image_path)))
+    
+    # 清理临时文件
+    os.unlink(image_path)
+
+# # 游戏中心帮助
+GameHelp = on_regex(pattern=r'^游戏帮助$',priority=1)
+@GameHelp.handle()
+async def handle_game_help(bot: Bot, event: GroupMessageEvent, state: T_State):
+    if not await check_command_enabled(bot, event, "游戏帮助"):
+        return
+    
+    # 发送处理提示
+    processing_msg = await bot.send(event=event, message="正在生成游戏帮助信息，请稍候...")
+    
+    # 生成帮助页面内容
+    html_content = render_game_help()
+    
+    # 转换为图片
+    # image_path = html_to_image(html_content)
+    image_path = await generate_html_screenshot(html_content)
+    
+    # 发送图片
+    await GameHelp.finish(MessageSegment.image(path_to_base64(image_path)))
     
     # 清理临时文件
     os.unlink(image_path)
